Упрощещенство

+++ handlers/default_handlers/topic_handler.py
@@ -1,41 +1,42 @@
 import traceback
 from pydoc_data.topics import topics
-
 from aiogram import F, Router
 from aiogram.fsm.context import FSMContext
 from aiogram.types import CallbackQuery, Message
 from aiogram.utils.keyboard import InlineKeyboardBuilder
-from loader import bot
 from config_data.config import CHAT_ID
-from handlers.default_handlers.adm_handlers import get_group_admins
 from loger.logger_helper import get_logger
-from keyboards.inline.inline_config import topic_kb
-from states.states_bot import Topic
+from keyboards.inline.inline_config import topic_chice
+from states.states_bot import Topic, Form, Builder
 from utils.utils_custom import delete_from_json, load_from_json, save_to_json
+from config_data.config import FILE_TOPIC, FILE_INFO
+from aiogram.filters import Command
 
+
+file_info = FILE_INFO
+file_topic = FILE_TOPIC
 router = Router()
 log = get_logger(__name__)
 chat_id = CHAT_ID
 
+text = "Выберете действие"
+async def topic_set(message: Message):
+    kb_done = await topic_chice()
+    await message.answer(text, reply_markup=kb_done)
+    await message.delete()
 
+
+
+
+@router.callback_query(Command("topic_menu"), Form.admin_true)
 @router.callback_query(F.data == "topic")
-async def topic_menu(callback: CallbackQuery, state: FSMContext):
-    """
-    Обрабатывает нажатие кнопки "Темы чата" в меню.
-
-    Отправляет пользователю клавиатуру с действиями, связанными с темами,
-    и устанавливает состояние машины состояний в Topic.menu.
-
-    Args:
-        callback (CallbackQuery): Объект обратного вызова, содержащий информацию о нажатой кнопке.
-        state (FSMContext): Контекст состояния для управления состояниями пользователя.
-    """
-    kb = await topic_kb()
-    await callback.message.answer("Выберите действие", reply_markup=kb)
+async def menu_topic(callback: CallbackQuery, state: FSMContext):
+    await topic_set(callback.message)
     await state.set_state(Topic.menu)
 
 
 @router.callback_query(F.data == "get_topic")
+@router.message(Command("topics"), Form.admin_true)
 async def set_topic(callback: CallbackQuery, state: FSMContext):
     """
     Обрабатывает нажатие кнопки "Добавить тему".
@@ -49,48 +50,40 @@ async def set_topic(callback: CallbackQuery, state: FSMContext):
         state (FSMContext): Контекст состояния для управления состояниями пользователя.
     """
     log.info("Пользователь выбрал 'get_topic'.")
-    await callback.message.answer("Введите: 'название темы'|'ссылка на тему'")
+    await callback.message.answer("Введите: 'название темы'|'ссылка на тему'\n"
+                                  "Для возвращения в меню темы введите 'Готово'")
     await state.set_state(Topic.menu)
+    await callback.message.delete()
 
 
-@router.message(Topic.menu)
+@router.message(Topic.menu, Topic.choice)
 async def get_topic(message: Message, state: FSMContext):
-    """
-    ОБНАВЛЕНО
-    Обрабатывает сообщение пользователя с информацией о теме.
-
-    Проверяет формат введенного текста, разбивает его на название темы
-    и ссылку, сохраняет информацию о теме в JSON-файл и уведомляет
-    администраторов о создании новой темы.
-
-    Args:
-        message (Message): Сообщение от пользователя, содержащее информацию о теме.
-        state (FSMContext): Контекст состояния для управления состояниями пользователя.
-    """
     text = message.text
-    lst = text.split('|')
-    if len(lst) != 2:
-        await message.answer("Неправильный формат. Попробуйте еще раз ввести 'название темы'|'ссылка на тему'")
-        return
-
-    try:
-        topic_name, topic_url = lst
-        data = {topic_name: {"url": topic_url}}
-        chat_info = await load_from_json("info_chats.json")
-        if "Темы чата" not in chat_info:
-            chat_info["Темы чата"] = data
-            await save_to_json(chat_info, "info_chats.json")
-        else:
-            data_topic = chat_info.get("Темы чата")
-            data_topic.update(data)
-            await save_to_json(chat_info, "info_chats.json")
-            log.debug(f"Чата с текущем ID: {topic_name} не существует")
-        await message.answer(f"Тема: {topic_name} принята")
-
-    except Exception as e:
-        log.error(f"Возникла ошибка: {e}\nТрассировка:\n{traceback.format_exc()}")
+    if text.lower() == "готово":
+        await topic_set(message)
         await state.set_state(Topic.menu)
-    await state.set_state(Topic.menu)
+    else:
+        lst = text.split('|')
+        if len(lst) != 2:
+            await message.answer("Неправильный формат. Попробуйте еще раз ввести 'название темы'|'ссылка на тему'")
+            return
+
+        try:
+            topic_name, topic_url = lst
+
+            data = {topic_name: {"url": topic_url}}
+            big_data = await load_from_json("topic.json")
+            if big_data.get(topic_name):
+                await message.reply(f"Тема: {topic_name} уже существует")
+                return
+            big_data.update(data)
+            await save_to_json(big_data, "topic.json")
+            await message.answer(f"Тема: {topic_name} принята")
+            await state.set_state(Topic.menu)
+
+        except Exception as e:
+            log.error(f"Возникла ошибка: {e}\nТрассировка:\n{traceback.format_exc()}")
+            await message.answer("Произошла ошибка при обработке вашей темы.")
 
 
 @router.callback_query(F.data == "view_topics")
@@ -106,25 +99,23 @@ async def view_topics(callback: CallbackQuery):
     """
 
     try:
-        data = await load_from_json("info_chats.json")
-        if data.get("Темы чата") is None:
+        data = await load_from_json("topic.json")
+        if data is None:
             await callback.message.answer("Темы не найдены.")
             return
-        topics_info = data.get("Темы чата")
-        log.debug(f"Для просмотра тем получены данные: {topics_info}")
+        log.debug(f"Для просмотра тем получены данные: {data}")
         response = "Список тем:\n"
-        for topic, info in topics_info.items():
-            log.debug(f"сообщение {info}")
-            response += f"{topics}: {info['url']}\n"
-        await callback.message.answer(response)
+        for topic in data:
+            log.debug(f"сообщение {topic}")
+            response += f"{topics}| url\n"
+            await callback.message.answer(response)
     except Exception as e:
         log.error(f"Возникла ошибка: {e}\nТрассировка:\n{traceback.format_exc()}")
 
 
-
-
-@router.callback_query(F.data == "delete_topic")
-async def choose_topic_for_deletion(callback: CallbackQuery):
+@router.callback_query(Topic.menu)
+@router.callback_query(F.data == "confirm")
+async def choose_topic_for_deletion(callback: CallbackQuery, state: FSMContext):
     """
     Обрабатывает нажатие кнопки "delete_topic".
 
@@ -134,7 +125,7 @@ async def choose_topic_for_deletion(callback: CallbackQuery):
     Args:
         callback (CallbackQuery): Объект обратного вызова, содержащий информацию о нажатой кнопке.
     """
-    log.info("Пользователь выбрал 'delete_topic'.")
+    log.info("Пользователь выбрал 'viewing_topic'.")
     topics = await load_from_json("topic.json")
     if not topics:
         await callback.message.answer("Темы не найдены.")
@@ -142,14 +133,21 @@ async def choose_topic_for_deletion(callback: CallbackQuery):
 
     builder = InlineKeyboardBuilder()
     for topic in topics.keys():
-        builder.button(text=topic, callback_data=f"confirm_delete_{topic}")
+        msg = ''
+        if callback.data == "deleted_topic":
+            msg = f"confirm_delete_{topic}"
+
+        elif callback.data == f"confirm":
+            msg = f"confirm_topic_{topic}"
+
+        builder.button(text=topic, callback_data=msg)
     builder.adjust(1)
 
     await callback.message.answer("Выберите тему:", reply_markup=builder.as_markup())
 
 
-@router.callback_query(F.data.startswith("confirm_delete_"))
-async def confirm_delete_topic(callback: CallbackQuery):
+@router.callback_query(F.data.startswith("confirm_"))
+async def confirm_delete_topic(callback: CallbackQuery, state: FSMContext):
     """
     Подтверждает удаление темы.
 
@@ -159,9 +157,25 @@ async def confirm_delete_topic(callback: CallbackQuery):
     Args:
         callback (CallbackQuery): Объект обратного вызова, содержащий информацию о нажатой кнопке.
     """
-    topic = callback.data[len("confirm_delete_"):]
-    success = await delete_from_json(topic, "topic.json")
-    if success:
-        await callback.message.answer(f"Тема '{topic}' успешно удалена.")
-    else:
-        await callback.message.answer(f"Не удалось удалить тему '{topic}'. Возможно, она не существует.")
+    try:
+        if callback.data.startswith("confirm_delete_"):
+            topic = callback.data[len("confirm_delete_"):]
+            log.debug(f"Удаление темы: {topic}")
+            success = await delete_from_json(topic, "topic.json")
+            if success:
+                await callback.message.answer(f"Тема '{topic}' успешно удалена.")
+                await topic_set(callback.message)
+        else:
+            topic = callback.data[len("confirm_topic_"):]
+            success = await load_from_json(topic, "topic.json")
+            id_topic = success.split('/')[-1]
+            await state.update_data(id_topic=id_topic)
+
+            await callback.message.answer(f"Тема {topic} выбрана")
+            await state.set_state(Builder.menu)
+        await callback.message.delete()
+    except Exception as e:
+        log.error(f"Возникла ошибка: {e}\nТрассировка:\n{traceback.format_exc()}")
+        await callback.message.answer("Не удалось выполнить действие")
+
+
